add IsFriendly, IsLegalMove; change isValidMove to IsAvailable, AvailableMoves to LegalMoves
previously had a method called isPotentialMove which was intended to be used inside of AvailableMoves, but also had a method called isValidMove. These naming of these methods made it confusing as to which was supposed to be used for a given circumstance. The new naming scheme implies a direct correlation between LegalMoves and isMoveLegal, leaving IsAvailable to indicated whether a Position is currently occupied or if it is available for a piece to move there.
closes issues #43